refactoring login system so it's a seperate page
