+++ lambda-notes/src/Components/EditNote/EditNote.js
@@ -44,7 +44,7 @@ class EditNote extends Component {
                             <textarea 
                             defaultValue={this.props.notes.notes[this.state.index].note}
                             onChange={this.handleInputChange} name="note" cols="99" rows="10" placeholder="Note Content"></textarea>
-                            <Link to="/" style={noDecoration}><button onClick={() => this.props.updateNote({title: this.state.title, note: this.state.note}, {index: this.state.index})}>Save</button></Link>
+                            <Link to="/home" style={noDecoration}><button onClick={() => this.props.updateNote({title: this.state.title, note: this.state.note}, {index: this.state.index})}>Save</button></Link>
                         </div>
                     </div>
                 )}
