+++ lambda-notes/src/Components/SideBar/SideBar.js
@@ -9,7 +9,7 @@ const SideBar = () => {
         <div className = "logo">
             <h1>Lambda Notes</h1>
         </div>
-        <Link to="/">View Your Notes</Link>
+        <Link to="/home">View Your Notes</Link>
         <Link to="/new">+ Create New Note</Link>
     </div>
     )
